move udpater into VencordSettings folder
