+++ src/components/index.ts
@@ -18,5 +18,4 @@
 
 export { default as PatchHelper } from "./PatchHelper";
 export { default as PluginSettings } from "./PluginSettings";
-export { default as Updater } from "./Updater";
 export { default as VencordSettings } from "./VencordSettings";
